Made Individual properties accessible for outer-class use.
Fit the cylinders inside the population, since it made more sense to. Additionally, sorted the cylinders into descending order of radius size, may change this to weight in the future, whilst almost setting the centre of the "heaviest" cylinder in the middle of the container.

Now, its to start the decoding process of the position strings.