Changes made within the decoding method in CylinderGroup so that it doesn't raise an exception when all cylinders, apart from the first, cannot be placed feasibly.
The previous commits concern is resolved by trimming all the cylinders from the start of the cylinders list that have an initial weight that exceeds  the maximum.

Additionally, a custom exception was set in case the maximum weight is too small for the problem.

Other things like adding a title to the visualisation was added, alongside a couple of other changes.