UNSTABLE COMMIT
Instead of discarding cylinders, a bin packing approach is being used. This is enables a more practical approach to pack all the cylinders in different containers while adhering to all the key constraints.

Additionally, have added some more evoluationary processes like single point crossover and a mutation suited for this problem.

THIS IS AN UNSTABLE COMMIT AND SO DO NOT RUN, IT'S JUST TO SAVE PROGRESS.